ndnd: Fixing a big bug with detecting TLV/NDNb format
There was error in the implementation due to erroneous thinking that
`buf' variable points to the same place as face->inbuf->buf.  The
variable actually points to start of the buffer where the last-read-data
was put.  As a result of the error, the decision could be made on a random
place of the packet, resulting in very unpredictable errors with TCP
faces.